small updates to jmePlots_run3.py

+++ scripts/jmePlots_run3_trkIter2Test.sh
@@ -12,7 +12,9 @@ jmePlots_run3.py -o ${OUTDIR} -l QCD_Pt_170to300 -e pdf png root -i \
  ${JMEANA_BASE}/analysis_output_run3_v02/outputs_postHarvesting/HLT_iter2RegionalPtSeed10p0/Run3Winter20_QCD_Pt_170to300_14TeV.root:'Reg 10.0':2 \
  ${JMEANA_BASE}/analysis_output_run3_v02/outputs_postHarvesting/HLT_iter2GlobalPtSeed0p9/Run3Winter20_QCD_Pt_170to300_14TeV.root:'Glo 0.9':4
 
-tar cfz ${OUTDIR%%/*}.tar.gz ${OUTDIR%%/*}
-rm -rf ${OUTDIR%%/*}
+if [ ! -f ${OUTDIR%%/*}.tar.gz ]; then
+  tar cfz ${OUTDIR%%/*}.tar.gz ${OUTDIR%%/*}
+  rm -rf ${OUTDIR%%/*}
+fi
 
 unset -v OUTDIR
